+++ CHANGELOG.md
@@ -11,3 +11,5 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - [x] feature: generalize `path_file` argument type, from `&str` to `AsRef<Path>`
 - [x] feature: allow only regular file for hashing with `sum_file`
 - [x] enhancement: add additional entropy to filename to fix race condition issue on parallel tests launch
+- [x] feature: implement CLI application
+- [x] documentation: add basic and advanced usage examples